fix proof mode display
fixes a bug where symbols would display as proof selected when not
in proof mode. Moves the cutBorder class to its own file
