+++ index.html
@@ -4,6 +4,7 @@
 	<meta charset="utf-8">
 	<script type="module" src="src/main.js"></script>
 	<script type="module" src="src/renderer.js"></script>
+	<script type="module" src="src/cutBorder.js"></script>
 	<script type="module" src="src/cut.js"></script>
 	<script type="module" src="src/userInput.js"></script>
 	<script type="module" src="src/symbol.js"></script>
